Refactor redux maker
