+++ frontend-admin/src/containers/Users/reducers.js
@@ -1,15 +1,5 @@
 import userModel from './actions';
 
-const initialState = {
-  listData: {
-    count: 0,
-    results: []
-  },
-  errors: {},
-  loading: false,
-  reqStatus: {}
-};
-
-const reducer = userModel.reducer(initialState);
+const reducer = userModel.reducer();
 
 export default reducer;
